# ggml : group mul_mat_id rows by matrix (cpu only)