An experiment to see if we can improve mixtral pp performance by grouping rows by matrix. CPU only.

With `llama-bench` the performance is mostly the same, but I think this is because it always uses the same token, so the same experts are selected. With `perplexity` it seems to reduce times by half.

We could do more grouping by processing all the experts in a single `ggml_mul_mat_id`.