@ggerganov 2023-12-15
With the `Q4_0` model, on AMD Ryzen 9 5950X 16-Core Processor:

```bash
make -j perplexity
time ./perplexity \
  -m ./models/mixtral-8x7b-32k/ggml-model-q4_0.gguf \
  -f ./tmp/mnt/models/wikitext/wikitext-2-raw/wiki.test.raw \
  --chunks 3
```

- `master`

```
system_info: n_threads = 16 / 32 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 0 | SSE3 = 1 | SSSE3 = 1 | VSX = 0 | 
perplexity: tokenizing the input ..
perplexity: tokenization took 682.277 ms
perplexity: calculating perplexity over 3 chunks, batch_size=512
perplexity: 76.32 seconds per pass - ETA 3.80 minutes
[1]3.3870,[2]3.9368,[3]4.6433,
Final estimate: PPL = 4.6433 +/- 0.38007

llama_print_timings:        load time =    2235.27 ms
llama_print_timings:      sample time =       0.00 ms /     1 runs   (    0.00 ms per token,      inf tokens per second)
llama_print_timings: prompt eval time =  231023.06 ms /  1536 tokens (  150.41 ms per token,     6.65 tokens per second)
llama_print_timings:        eval time =       0.00 ms /     1 runs   (    0.00 ms per token,      inf tokens per second)
llama_print_timings:       total time =  231831.13 ms

real	3m55,269s
user	61m43,347s
sys	0m3,304s
```

- PR

```
system_info: n_threads = 16 / 32 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 0 | SSE3 = 1 | SSSE3 = 1 | VSX = 0 | 
perplexity: tokenizing the input ..
perplexity: tokenization took 690.123 ms
perplexity: calculating perplexity over 3 chunks, batch_size=512
perplexity: 20.26 seconds per pass - ETA 1.00 minutes
[1]3.3870,[2]3.9368,[3]4.6433,
Final estimate: PPL = 4.6433 +/- 0.38007

llama_print_timings:        load time =    2264.67 ms
llama_print_timings:      sample time =       0.00 ms /     1 runs   (    0.00 ms per token,      inf tokens per second)
llama_print_timings: prompt eval time =   60974.41 ms /  1536 tokens (   39.70 ms per token,    25.19 tokens per second)
llama_print_timings:        eval time =       0.00 ms /     1 runs   (    0.00 ms per token,      inf tokens per second)
llama_print_timings:       total time =   61778.02 ms

real	1m5,256s
user	16m22,577s
sys	0m3,304s
```

Looks like this PR is more than 3x faster on this particular machine with CPU

Note:
- OpenBLAS is disabled
- this is the old model with the wrong `freq_base = 1e4` (don't have the new file on this machine), so the ppl values are irrelevant