@wtarreau 2023-12-16
The gain is indeed impressive, before (b1641) and after (b1644), on an Ampere Altra 80 cores with mixtral-8x7b:

b1641:
```
llama_print_timings:        load time =    2734.08 ms
llama_print_timings:      sample time =    3258.74 ms /  1366 runs   (    2.39 ms per token,   419.18 tokens per second)
llama_print_timings: prompt eval time =   87993.18 ms /  1778 tokens (   49.49 ms per token,    20.21 tokens per second)
llama_print_timings:        eval time =  145351.73 ms /  1365 runs   (  106.48 ms per token,     9.39 tokens per second)
llama_print_timings:       total time =  238797.08 ms
Log end

real    4m2.475s
user    304m55.246s
sys     0m40.517s
```

b1644:
```
llama_print_timings:        load time =    2711.56 ms
llama_print_timings:      sample time =    3247.65 ms /  1366 runs   (    2.38 ms per token,   420.61 tokens per second)
llama_print_timings: prompt eval time =   53843.62 ms /  1778 tokens (   30.28 ms per token,    33.02 tokens per second)
llama_print_timings:        eval time =  131554.21 ms /  1365 runs   (   96.38 ms per token,    10.38 tokens per second)
llama_print_timings:       total time =  188936.37 ms
Log end

real    3m12.606s
user    240m59.715s
sys     0m43.709s
```
It was visible to the naked eye that the eval time was 60% faster! Great job!